refs #370. Now, when publishing a workspace, the public data of the workspace it's published also. When a user links to a published workspace, it inherit the public variable values of the workspace.
Currently, the only value not published are password preferences. But all logic controlling what varaibles are public or not it's isolated in the VariableDef.has_public_value() method.

The above method will be reimplement when the template allows expresing for each varaible, if it should be exported it's value in the mashup, or not.

--HG--
extra : convert_revision : svn%3A2a98c5bb-d66c-48ca-a96c-db028a6c31f4/ezweb_platform/src/trunk%401889