Fix proxy problem in Django 1.3
In Django 1.3 when the raw_post_data attribute is read, the
socket is queried and if no Content-Length header is present,
that call blocks.

It works with the runserver command because there is some extra
logic to avoid this case and the socket is wrapped in a class
that checks if there is something on the wire.

The fix is simple: don't read the raw_post_data attribute if
the request is not a POST (or PUT).

--HG--
extra : rebase_source : 3e90ea8ec1ee8918707eaecafa4f57ebcee07d47