Avoid save None values into the DB when linking a workspace.
When a workspace is linked to another user, all the variables are intialized
with the values the creator of the workspace had for that variables. When
the creator doesn't have a variable initialized, then it uses the default
value of the variable to initilize the new one. But that default value
may be ''None'', this changeset avoids initialize the new variable in that case.