Fix small problem in the UploadWindowMenu dialog.
This dialog doesn't work if you open the upload dialog and then you open
the file browser and once added to the list of files to upload, cancels
the upload. Open one time again the upload dialog, open the file browser
and make the same selection. After this step, the dialog doesn't add the
selection to the list.