Add widget builder
A series of scripts to download code from GitHub and compile the widget. There are issues with the version of libxml used by most widgets which means that it needs to re-compile sources when running npm install, otherwise the Docker script is just a vanilla NodeJS with a few extras.

The listing script assume a linux system with xsltproc and sed to list details of all Widgets.